refactor logic to hook and create game engine
