+++ src/engine/index.tsx
@@ -1,0 +1,11 @@
+import React, { FunctionComponent } from "react";
+
+// type Props = {
+//   // children: React.ReactChildren;
+// };
+
+const GameEngine: FunctionComponent = (props) => {
+  return <>{props.children}</>;
+};
+
+export default GameEngine;
