+++ src/views/GameRouter/components/GameTable/index.tsx
