+++ src/views/GameRouter/components/GameTable/styles.ts
