+++ src/views/GameRouter/components/TablePlayer/style.ts
